requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.16
langchain-experimental>=0.0.40
langchain-core>=0.1.18
langchain-text-splitters>=0.0.1
langchain-chroma>=0.0.1
langchain-neo4j>=0.1.1
networkx>=3.1
openai>=1.3.0
numpy>=1.23.0
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
Levenshtein>=0.21.0
neo4j>=5.13.0
matplotlib>=3.7.0
python-louvain>=0.16
scikit-learn>=1.3.0